+++ src/ui/map_tests/map_zoom.test.ts
@@ -1,38 +1,8 @@
-import {Map, MapOptions} from '../map';
-import {createMap, beforeMapTest, sleep, createStyle, createStyleSource} from '../../util/test/util';
-import {LngLat} from '../../geo/lng_lat';
-import {Tile} from '../../source/tile';
-import {OverscaledTileID} from '../../source/tile_id';
-import {Event as EventedEvent, ErrorEvent} from '../../util/evented';
-import simulate from '../../../test/unit/lib/simulate_interaction';
-import {fixedLngLat, fixedNum} from '../../../test/unit/lib/fixed';
-import {GeoJSONSourceSpecification, LayerSpecification, SourceSpecification, StyleSpecification} from '@maplibre/maplibre-gl-style-spec';
-import {RequestTransformFunction} from '../../util/request_manager';
-import {extend} from '../../util/util';
-import {LngLatBoundsLike} from '../../geo/lng_lat_bounds';
-import {IControl} from '../control/control';
-import {EvaluationParameters} from '../../style/evaluation_parameters';
-import {fakeServer, FakeServer} from 'nise';
-import {CameraOptions} from '../camera';
-import {Terrain} from '../../render/terrain';
-import {mercatorZfromAltitude} from '../../geo/mercator_coordinate';
-import {Transform} from '../../geo/transform';
-import {StyleImageInterface} from '../../style/style_image';
-import {Style} from '../../style/style';
-import {MapSourceDataEvent} from '../events';
-import {config} from '../../util/config';
-import {MessageType} from '../../util/actor_messages';
-
-let server: FakeServer;
+import {createMap, beforeMapTest} from '../../util/test/util';
 
 beforeEach(() => {
     beforeMapTest();
     global.fetch = null;
-    server = fakeServer.create();
-});
-
-afterEach(() => {
-    server.restore();
 });
 
 test('#setMinZoom', () => {
@@ -55,7 +25,6 @@ test('#getMinZoom', () => {
     map.setMinZoom(10);
     expect(map.getMinZoom()).toBe(10);
 });
-
 
 test('ignore minZooms over maxZoom', () => {
     const map = createMap({zoom: 2, maxZoom: 5});
